Fix paths and add path sanity checks.
This change factors out the paths that the tests rely on from inlines to constants and adds a set of tests that check to see if the paths they refer to exist or do not exist, as desired. This should make future issues of configuration/distribution significantly easier to catch.